feat: code refactor for auth
