--- rhAPI/src/main/java/mg/itu/rh/auth/AuthService.java
@@ -1,46 +1,0 @@
-package mg.itu.rh.auth;
-
-import mg.itu.rh.auth.exception.InvalidCredentialsException;
-import mg.itu.rh.entity.interne.Contrat;
-import mg.itu.rh.entity.talent.Talent;
-import mg.itu.rh.repository.talent.TalentRepository;
-import mg.itu.rh.service.interne.ContratService;
-import org.springframework.stereotype.Service;
-
-import java.util.Optional;
-
-@Service
-public class AuthService {
-    private final TalentRepository talentRepository;
-    private final ContratService contratService;
-
-    public AuthService(TalentRepository talentRepository, ContratService contratService) {
-        this.talentRepository = talentRepository;
-        this.contratService = contratService;
-    }
-
-    public LoginResponse findByEmailAndPassword(LoginRequest authDTO) throws InvalidCredentialsException {
-        String email = authDTO.getEmail(), pwd = authDTO.getPassword();
-        Talent t = talentRepository.findByEmailAndPassword(email, pwd)
-                .orElseThrow(() -> new InvalidCredentialsException());
-
-        LoginResponse responseDTO = new LoginResponse(t);
-        if (t.isMfaEnabled()) {
-            responseDTO.setMfaRequired(true);
-            return responseDTO;
-        }
-
-        Optional<Contrat> c = contratService.findActualContratByIdTalent(t.getIdTalent());
-        c.ifPresent(contrat -> responseDTO.setIdContrat(contrat.getIdContrat()));
-
-        return responseDTO;
-    }
-
-    public LoginResponse findByEmailAndPasswordBypassingMfa(Talent t) {
-        LoginResponse responseDTO = new LoginResponse(t);
-        Optional<Contrat> c = contratService.findActualContratByIdTalent(t.getIdTalent());
-        c.ifPresent(contrat -> responseDTO.setIdContrat(contrat.getIdContrat()));
-
-        return responseDTO;
-    }
-}
