--- rhAPI/src/main/java/mg/itu/rh/auth/LoginRequest.java
@@ -1,9 +1,0 @@
-package mg.itu.rh.auth;
-
-import lombok.Data;
-
-@Data
-public class LoginRequest {
-    private String email;
-    private String password;
-}
