--- rhAPI/src/main/java/mg/itu/rh/auth/MFARequest.java
@@ -1,10 +1,0 @@
-package mg.itu.rh.auth;
-
-import lombok.Data;
-
-@Data
-public class MFARequest {
-    private String email;
-    private int code;
-    private String scratchCode;
-}
