+++ rhAPI/src/main/java/mg/itu/rh/auth/dto/LoginRequest.java
@@ -1,0 +1,14 @@
+package mg.itu.rh.auth.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Data;
+
+@Data
+public class LoginRequest {
+    @Email(message = "Invalid email format")
+    private String email;
+
+    @NotBlank(message = "Password is required")
+    private String password;
+}
