+++ rhAPI/src/main/java/mg/itu/rh/auth/exception/InvalidMfaException.java
@@ -6,6 +6,6 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 @ResponseStatus(HttpStatus.UNAUTHORIZED)
 public class InvalidMfaException extends RuntimeException {
     public InvalidMfaException() {
-        super("Invalid MFA code");
+        super("Invalid MFA code or scratch code");
     }
 }
